refactor: make validate its own module
